Hide transport error types (except BadRequest/Unexpected)
`BadRequestError` and `UnexpectedError` will be part of the public interface.
